refactor: Split serviceslist into smaller components for better organization and maintainability
